feat: 카테고리 컨트롤러 통합테스트 작성
